move updateclient to actionhandler
