+++ app/src/action_handler/actions/ibc_action/msg/update_client.rs
@@ -1,16 +1,26 @@
 use std::sync::Arc;
 
-use anyhow::Result;
+use anyhow::{Context, Result};
 use async_trait::async_trait;
+use ibc::core::ics02_client::client_state::ClientState;
 use ibc::core::ics02_client::msgs::update_client::MsgUpdateClient;
+use ibc::core::ics24_host::identifier::ClientId;
+use penumbra_chain::StateReadExt as _;
 use penumbra_storage::{StateRead, StateWrite};
 use penumbra_transaction::Transaction;
 
+use ibc::clients::ics07_tendermint::client_state::ClientState as TendermintClientState;
+use ibc::clients::ics07_tendermint::consensus_state::ConsensusState as TendermintConsensusState;
+use ibc::clients::ics07_tendermint::header::Header as TendermintHeader;
+use tendermint::validator;
+use tendermint_light_client_verifier::types::{TrustedBlockState, UntrustedBlockState};
+use tendermint_light_client_verifier::{ProdVerifier, Verdict, Verifier};
+
 use crate::action_handler::ActionHandler;
-use crate::ibc::component::client::stateless::update_client::header_is_tendermint;
-use crate::ibc::component::client::{
-    stateful::update_client::UpdateClientCheck, Ics2ClientExt as _,
-};
+use crate::ibc::client::ics02_validation;
+use crate::ibc::component::client::Ics2ClientExt as _;
+use crate::ibc::component::client::{StateReadExt as _, StateWriteExt as _};
+use crate::ibc::event;
 
 #[async_trait]
 impl ActionHandler for MsgUpdateClient {
@@ -30,14 +40,227 @@ impl ActionHandler for MsgUpdateClient {
         // to Ok(()) rather than erroring to avoid having two "racing" relay
         // transactions fail just because they both contain the same client
         // update.
-        if !state.update_is_already_committed(&self).await? {
+        if !update_is_already_committed(&mut state, &self).await? {
             tracing::debug!(msg = ?self);
-            state.validate(self).await?;
-            state.execute_update_client(self).await?;
+
+            let client_state = client_is_present(&mut state, self).await?;
+
+            client_is_not_frozen(&client_state)?;
+            client_is_not_expired(&mut state, &self.client_id, &client_state).await?;
+
+            let trusted_client_state = client_state;
+
+            let untrusted_header = ics02_validation::get_tendermint_header(self.header.clone())?;
+
+            header_revision_matches_client_state(&trusted_client_state, &untrusted_header)?;
+            header_height_is_consistent(&untrusted_header)?;
+
+            // The (still untrusted) header uses the `trusted_height` field to
+            // specify the trusted anchor data it is extending.
+            let trusted_height = untrusted_header.trusted_height;
+
+            // We use the specified trusted height to query the trusted
+            // consensus state the update extends.
+            let last_trusted_consensus_state = state
+                .get_verified_consensus_state(trusted_height, self.client_id.clone())
+                .await?;
+
+            let last_trusted_consensus_state = last_trusted_consensus_state;
+
+            // We also have to convert from an IBC height, which has two
+            // components, to a Tendermint height, which has only one.
+            let trusted_height = trusted_height
+                .revision_height()
+                .try_into()
+                .context("invalid header height")?;
+
+            let trusted_validator_set =
+                verify_header_validator_set(&untrusted_header, &last_trusted_consensus_state)?;
+
+            // Now we build the trusted and untrusted states to feed to the Tendermint light client.
+
+            let trusted_state = TrustedBlockState {
+                // TODO(erwan): do we need an additional check on `chain_id`
+                chain_id: &trusted_client_state.chain_id.clone().into(),
+                header_time: last_trusted_consensus_state.timestamp,
+                height: trusted_height,
+                next_validators: trusted_validator_set,
+                next_validators_hash: last_trusted_consensus_state.next_validators_hash,
+            };
+
+            let untrusted_state = UntrustedBlockState {
+                signed_header: &untrusted_header.signed_header,
+                validators: &untrusted_header.validator_set,
+                next_validators: None, // TODO: do we need this?
+            };
+
+            let options = trusted_client_state.as_light_client_options()?;
+            let verifier = ProdVerifier::default();
+
+            let verdict = verifier.verify(
+                untrusted_state,
+                trusted_state,
+                &options,
+                state.get_block_timestamp().await?,
+            );
+
+            match verdict {
+                Verdict::Success => Ok(()),
+                Verdict::NotEnoughTrust(voting_power_tally) => Err(anyhow::anyhow!(
+                    "not enough trust, voting power tally: {:?}",
+                    voting_power_tally
+                )),
+                Verdict::Invalid(detail) => Err(anyhow::anyhow!(
+                    "could not verify tendermint header: invalid: {:?}",
+                    detail
+                )),
+            }?;
+
+            let trusted_header = untrusted_header;
+
+            // get the latest client state
+            let client_state = state.get_client_state(&self.client_id).await.unwrap();
+
+            let (next_tm_client_state, next_tm_consensus_state) = state
+                .next_tendermint_state(
+                    self.client_id.clone(),
+                    client_state.clone(),
+                    trusted_header.clone(),
+                )
+                .await;
+
+            // store the updated client and consensus states
+            state.put_client(&self.client_id, next_tm_client_state);
+            state
+                .put_verified_consensus_state(
+                    trusted_header.height(),
+                    self.client_id.clone(),
+                    next_tm_consensus_state,
+                )
+                .await?;
+
+            state.record(event::update_client(
+                self.client_id.clone(),
+                client_state,
+                trusted_header,
+            ));
+
+            return Ok(());
         } else {
             tracing::debug!("skipping duplicate update");
         }
 
         Ok(())
+    }
+}
+
+fn header_is_tendermint(msg: &MsgUpdateClient) -> anyhow::Result<()> {
+    if ics02_validation::is_tendermint_header_state(&msg.header) {
+        Ok(())
+    } else {
+        Err(anyhow::anyhow!("MsgUpdateClient: not a tendermint header"))
+    }
+}
+
+async fn update_is_already_committed<S: StateRead>(
+    state: S,
+    msg: &MsgUpdateClient,
+) -> anyhow::Result<bool> {
+    let untrusted_header = ics02_validation::get_tendermint_header(msg.header.clone())?;
+    let client_id = msg.client_id.clone();
+
+    // check if we already have a consensus state for this height, if we do, check that it is
+    // the same as this update, if it is, return early.
+    let untrusted_consensus_state = TendermintConsensusState::from(untrusted_header.clone());
+    if let Ok(stored_consensus_state) = state
+        .get_verified_consensus_state(untrusted_header.height(), client_id.clone())
+        .await
+    {
+        let stored_tm_consensus_state = stored_consensus_state;
+
+        Ok(stored_tm_consensus_state == untrusted_consensus_state)
+    } else {
+        // If we don't have a consensus state for this height for
+        // whatever reason (either missing or a DB error), we don't
+        // consider it an error, it's just not already committed.
+        Ok(false)
+    }
+}
+
+async fn client_is_not_expired<S: StateRead>(
+    state: S,
+    client_id: &ClientId,
+    client_state: &TendermintClientState,
+) -> anyhow::Result<()> {
+    let latest_consensus_state = state
+        .get_verified_consensus_state(client_state.latest_height(), client_id.clone())
+        .await?;
+
+    // TODO(erwan): for now there is no casting that needs to happen because `get_verified_consensus_state` does not return an
+    // abstracted consensus state.
+    let latest_consensus_state_tm = latest_consensus_state;
+
+    let now = state.get_block_timestamp().await?;
+    let time_elapsed = now.duration_since(latest_consensus_state_tm.timestamp)?;
+
+    if client_state.expired(time_elapsed) {
+        Err(anyhow::anyhow!("client is expired"))
+    } else {
+        Ok(())
+    }
+}
+
+async fn client_is_present<S: StateRead>(
+    state: S,
+    msg: &MsgUpdateClient,
+) -> anyhow::Result<TendermintClientState> {
+    state.get_client_type(&msg.client_id).await?;
+
+    state.get_client_state(&msg.client_id).await
+}
+
+fn client_is_not_frozen(client: &TendermintClientState) -> anyhow::Result<()> {
+    if client.is_frozen() {
+        Err(anyhow::anyhow!("client is frozen"))
+    } else {
+        Ok(())
+    }
+}
+
+fn header_revision_matches_client_state(
+    trusted_client_state: &TendermintClientState,
+    untrusted_header: &TendermintHeader,
+) -> anyhow::Result<()> {
+    if untrusted_header.height().revision_number() != trusted_client_state.chain_id.version() {
+        Err(anyhow::anyhow!(
+            "client update revision number does not match client state"
+        ))
+    } else {
+        Ok(())
+    }
+}
+
+fn header_height_is_consistent(untrusted_header: &TendermintHeader) -> anyhow::Result<()> {
+    if untrusted_header.height() <= untrusted_header.trusted_height {
+        Err(anyhow::anyhow!(
+            "client update height is not greater than trusted height"
+        ))
+    } else {
+        Ok(())
+    }
+}
+
+fn verify_header_validator_set<'h>(
+    untrusted_header: &'h TendermintHeader,
+    last_trusted_consensus_state: &TendermintConsensusState,
+) -> anyhow::Result<&'h validator::Set> {
+    if untrusted_header.trusted_validator_set.hash()
+        != last_trusted_consensus_state.next_validators_hash
+    {
+        Err(anyhow::anyhow!(
+            "client update validator set hash does not match trusted consensus state"
+        ))
+    } else {
+        Ok(&untrusted_header.trusted_validator_set)
     }
 }
